rename examples/agent_mesh_maze -> examples/agent_maze
"mesh" is de-emphasized now that the demo uses Decentralized/Centralized as the
arm names, so the shorter agent_maze is clearer. Updated all module paths
(imports, the `python -m examples.agent_maze.run` command in README/run.py),
and the pyproject per-file-ignore glob. Imports + replay path + all CI gates
(ruff, ty, codespell) verified green under the new name.
