+++ examples/agent_maze/web/data.js
