+++ examples/agent_maze/world.py
